+++ app/models/site.rb
@@ -225,11 +225,7 @@ class Site < ActiveRecord::Base
   end
 
   def set_rails_tz
-    tzInfo_id = TimeZoneHelper.to_identifier(self.tzinfo_tz)
-    rails_tz_string = TimeZoneHelper.tzinfo_to_ruby(tzInfo_id)
-    unless rails_tz_string.blank?
-      self.rails_tz = rails_tz_string
-    end
+    TimeZoneHelper.set_rails_tz(self)
   end
 
 end